move actor, dispatch, monitor to v2
